feat(ui): Rename sorted table to data table and show no data message with filter
