+++ ui/src/app/shared/shared.module.ts
@@ -57,7 +57,7 @@ import {RequirementsFormComponent} from './requirements/form/requirements.form.c
 import {RequirementsListComponent} from './requirements/list/requirements.list.component';
 import {SharedService} from './shared.service';
 import {StatusIconComponent} from './status/status.component';
-import {SortedTableComponent} from './table/sorted-table.component';
+import {DataTableComponent} from './table/data-table.component';
 import {ToastService} from './toast/ToastService';
 import {TokenListComponent} from './token/list/token.list.component';
 import {UsageApplicationsComponent} from './usage/applications/usage.applications.component';
@@ -176,7 +176,7 @@ import {ZoneComponent} from './zone/zone.component';
         SpanColoredComponent,
         VCSStrategyComponent,
         FavoriteCardsComponent,
-        SortedTableComponent
+        DataTableComponent
     ],
     entryComponents: [SpanColoredComponent],
     providers: [
@@ -272,7 +272,7 @@ import {ZoneComponent} from './zone/zone.component';
         SpanColoredComponent,
         VCSStrategyComponent,
         FavoriteCardsComponent,
-        SortedTableComponent
+        DataTableComponent
     ]
 })
 export class SharedModule {
