+++ ui/src/app/shared/table/data-table.html
@@ -25,14 +25,14 @@
             </td>
         </tr>
     </tbody>
-    <tbody *ngIf="!loading && data.length == 0">
+    <tbody *ngIf="!loading && dataForCurrentPage.length == 0">
         <tr>
             <td class="center" [colSpan]="columns.length">
                 <div>{{ 'common_no_data' | translate }}</div>
             </td>
         </tr>
     </tbody>
-    <tbody *ngIf="!loading && data.length > 0">
+    <tbody *ngIf="!loading && dataForCurrentPage.length > 0">
         <tr *ngFor="let d of dataForCurrentPage">
             <ng-container *ngFor="let c of columns" [ngSwitch]="c.type">
                 <td *ngSwitchCase="'icon'" class="border">
